requirements.txt
Django==3.2.6
psycopg2==2.9.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
